+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   * On very fast Therion executions, therion.log file content not being included in Therion run output.
   * Bézier created from straight lines are too tight.
   * Status bar info for selected elements not internationalized.
+  * Current PLA type at PLA type selection overlay with too small left and bottom margins.
 * Infrastructure maintenance:
   * Enabling github created flatpak generation.
   * New "New version available" page.